agents/makoto/demos/line_bot_echo/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
line-bot-sdk==3.6.0
python-dotenv==1.0.1
